+++ providers/common/compat/src/airflow/providers/common/compat/notifier/__init__.py
@@ -22,9 +22,9 @@ from typing import TYPE_CHECKING
 from airflow.providers.common.compat.version_compat import AIRFLOW_V_3_0_PLUS
 
 if TYPE_CHECKING:
-    from airflow.sdk.definitions.notifier import BaseNotifier
+    from airflow.sdk.bases.notifier import BaseNotifier
 elif AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk.definitions.notifier import BaseNotifier
+    from airflow.sdk.bases.notifier import BaseNotifier
 else:
     from airflow.notifications.basenotifier import BaseNotifier
