+++ providers/common/io/src/airflow/providers/common/io/xcom/backend.py
@@ -37,7 +37,7 @@ if TYPE_CHECKING:
     from airflow.sdk.execution_time.comms import XComResult
 
 if AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk.execution_time.xcom import BaseXCom
+    from airflow.sdk.bases.xcom import BaseXCom
 else:
     from airflow.models.xcom import BaseXCom  # type: ignore[no-redef]
