+++ task-sdk/src/airflow/sdk/definitions/_internal/abstractoperator.py
@@ -41,8 +41,8 @@ from airflow.utils.weight_rule import WeightRule
 if TYPE_CHECKING:
     import jinja2
 
-    from airflow.sdk.definitions.baseoperator import BaseOperator
-    from airflow.sdk.definitions.baseoperatorlink import BaseOperatorLink
+    from airflow.sdk.bases.baseoperator import BaseOperator
+    from airflow.sdk.bases.operatorlink import BaseOperatorLink
     from airflow.sdk.definitions.context import Context
     from airflow.sdk.definitions.dag import DAG
     from airflow.sdk.definitions.mappedoperator import MappedOperator
