+++ task-sdk/src/airflow/sdk/definitions/_internal/mixins.py
@@ -22,8 +22,8 @@ from collections.abc import Iterable, Sequence
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
+    from airflow.sdk.bases.baseoperator import BaseOperator
     from airflow.sdk.definitions._internal.abstractoperator import Operator
-    from airflow.sdk.definitions.baseoperator import BaseOperator
     from airflow.sdk.definitions.context import Context
     from airflow.sdk.definitions.edges import EdgeModifier
