+++ task-sdk/tests/task_sdk/bases/test_sensor.py
@@ -34,8 +34,8 @@ from airflow.exceptions import (
 )
 from airflow.models.trigger import TriggerFailureReason
 from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.sdk.bases.sensor import BaseSensorOperator, PokeReturnValue, poke_mode_only
 from airflow.sdk.definitions.dag import DAG
-from airflow.sdk.definitions.sensors.base import BaseSensorOperator, PokeReturnValue, poke_mode_only
 from airflow.sdk.execution_time.comms import RescheduleTask, TaskRescheduleStartDate
 from airflow.utils import timezone
 from airflow.utils.state import State
