+++ task-sdk/tests/task_sdk/dags/super_basic.py
@@ -17,7 +17,7 @@
 
 from __future__ import annotations
 
-from airflow.sdk.definitions.baseoperator import BaseOperator
+from airflow.sdk.bases.baseoperator import BaseOperator
 from airflow.sdk.definitions.dag import dag
