+++ task-sdk/tests/task_sdk/definitions/test_mixins.py
@@ -22,7 +22,7 @@ import itertools
 import pytest
 
 from airflow.decorators import setup, task, teardown
-from airflow.sdk.definitions.baseoperator import BaseOperator
+from airflow.sdk.bases.baseoperator import BaseOperator
 from airflow.sdk.definitions.dag import DAG
